Corrections in ppa's

+++ extras/development-apps.sh
@@ -20,7 +20,7 @@
 
 # Android Studio
 # --------------------------------------
-sudo apt install curl
+sudo apt install curl -y -q
 
 
 # Android Studio
@@ -30,8 +30,10 @@ sudo snap install android-studio --classic
 
 # Atom
 # --------------------------------------
-wget -qO - https://packagecloud.io/AtomEditor/atom/gpgkey | sudo gpg --dearmor -o /usr/share/keyrings/atom-keyring.gpg
-echo "deb [arch=amd64 signed-by=/usr/share/keyrings/atom-keyring.gpg] https://packagecloud.io/AtomEditor/atom/any/ any main" | sudo tee /etc/apt/sources.list.d/atom.list
+if [ ! -f "/usr/share/keyrings/atom-keyring.gpg" ]; then
+	wget -qO - https://packagecloud.io/AtomEditor/atom/gpgkey | sudo gpg --dearmor -o /usr/share/keyrings/atom-keyring.gpg
+	echo "deb [arch=amd64 signed-by=/usr/share/keyrings/atom-keyring.gpg] https://packagecloud.io/AtomEditor/atom/any/ any main" | sudo tee /etc/apt/sources.list.d/atom.list
+fi
 
 sudo apt update
 sudo apt install atom -y
@@ -52,8 +54,10 @@ git config --global user.name "Tiago Lucas Flach"
 
 ## GitHub CLI
 # --------------------------------------
-curl -fsSL https://cli.github.com/packages/githubcli-archive-keyring.gpg | sudo dd of=/usr/share/keyrings/githubcli-archive-keyring.gpg
-echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/githubcli-archive-keyring.gpg] https://cli.github.com/packages stable main" | sudo tee /etc/apt/sources.list.d/github-cli.list > /dev/null
+if [ ! -f "/usr/share/keyrings/githubcli-archive-keyring.gpg" ]; then
+	curl -fsSL https://cli.github.com/packages/githubcli-archive-keyring.gpg | sudo dd of=/usr/share/keyrings/githubcli-archive-keyring.gpg
+	echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/githubcli-archive-keyring.gpg] https://cli.github.com/packages stable main" | sudo tee /etc/apt/sources.list.d/github-cli.list > /dev/null
+fi
 
 sudo apt update
 sudo apt install gh
@@ -85,8 +89,10 @@ sudo snap install phpstorm --classic
 
 # Subllime Text
 # --------------------------------------
-wget -qO - https://download.sublimetext.com/sublimehq-pub.gpg | sudo gpg --dearmor -o /usr/share/keyrings/sublimetext-keyring.gpg
-echo "deb [arch=amd64 signed-by=/usr/share/keyrings/sublimetext-keyring.gpg] https://download.sublimetext.com/ apt/stable/" | sudo tee /etc/apt/sources.list.d/sublime-text.list
+if [ ! -f "/usr/share/keyrings/sublimetext-keyring.gpg" ]; then
+	wget -qO - https://download.sublimetext.com/sublimehq-pub.gpg | sudo gpg --dearmor -o /usr/share/keyrings/sublimetext-keyring.gpg
+	echo "deb [arch=amd64 signed-by=/usr/share/keyrings/sublimetext-keyring.gpg] https://download.sublimetext.com/ apt/stable/" | sudo tee /etc/apt/sources.list.d/sublime-text.list
+fi
 
 sudo apt update
 sudo apt install sublime-text -y
